make all hb related code lazy
